requirements.txt
pandas
fastapi
uvicorn[standard]
scikit-learn
numpy
python-dotenv
